Module bones_bevy_asset::prelude::bones::time

Time functionality for the Bones framework.

This is a slimmed down version of bevy_time.

bevy_time is licensed under MIT OR Apache-2.0.

Structs

Stopwatch

A Stopwatch is a struct that track elapsed time when started.

Time

A clock that tracks how much it has advanced (and how much real time has elapsed) since its previous update and since its creation.

Timer

Tracks elapsed time. Enters the finished state once duration is reached.

Enums

TimerMode

Specifies Timer behavior.