Trait bones_bevy_asset::BonesBevyAssetLoad

pub trait BonesBevyAssetLoad {
    fn load(
        &mut self,
        load_context: &mut LoadContext<'_>,
        dependencies: &mut Vec<AssetPath<'static>>
    ) { ... }
}

Trait implemented for types that may appear in the fields of a BonesBevyAsset and may need to perform aditional loading with the bevy load context.

Provided Methods

fn load(
&mut self,
load_context: &mut LoadContext<'_>,
dependencies: &mut Vec<AssetPath<'static>>
)

Allows the field to do any extra loading that it might need to do from the Bevy load context when the asset is loaded.